move out googlenet_spritemap() method from semantic_dict
