+++ src/utils.py
@@ -2,5 +2,19 @@
 define commonly used utilities for this repo
 """
 
-def test():
-    print("do you see me?")
+layer_spritemap_sizes = {
+    'mixed3a' : 16,
+    'mixed3b' : 21,
+    'mixed4a' : 22,
+    'mixed4b' : 22,
+    'mixed4c' : 22,
+    'mixed4d' : 22,
+    'mixed4e' : 28,
+    'mixed5a' : 28,
+  }
+
+def googlenet_spritemap(layer):
+  assert layer in layer_spritemap_sizes
+  size = layer_spritemap_sizes[layer]
+  url = "https://storage.googleapis.com/lucid-static/building-blocks/googlenet_spritemaps/sprite_%s_channel_alpha.jpeg" % layer
+  return size, url
